nuscenes_web_platform/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pydantic-settings>=2.6.0

# Align with demo_by_nuscenes
nuscenes-devkit==1.2.0
pyquaternion>=0.9.9
numpy>=1.23.0
matplotlib>=3.7.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
